requirements.txt
numpy>=1.24,<2.0
scipy>=1.10
pandas>=1.5
scikit-image>=0.21
opencv-python==4.7.0.72
tifffile>=2023.7
matplotlib>=3.7
seaborn>=0.12
mahotas>=1.4
PyYAML>=6.0
typer>=0.9
rich>=13.7
pytest>=7.4
torch>=2.2,<2.4
torchvision>=0.17,<0.20
torchdyn>=1.0.5
absl-py>=1.4
torchcfm>=0.3
